Реализован клиентский метод стягивания буфера с сервера. Рефакторинг старого кода.
 - класс _Dumper (тестовая пулялка) теперь использует на один уровень абстракции выше - SharedBufferReader и умеет выводить легкий бенчмарк стягивания буфера с сервера
 - клиенты теперь используют абстрактный интерфейс к разделяемой памяти
 - внутреннее устройство областей памяти (Internal) теперь PIMPL
 - метод getBuffer теперь называется getRawBuffer, т.к. он возвращает сырой указатель на данные, что не очень ок
 - класс `LowLevelBufferHandler` теперь имеет метод parseBuffer как результат функциональной декомпозиции. Легче читать
 - также класс `LowLevelBufferHandler` теперь умеет читать буфер напрямую в контейнер типа, что показало увеличение быстродействия в 10 раз за счет устранения двойного копирования (теперь копирование одно). Метод называется `getBuffer` и реализован на шаблонах.
 - теперь класс `LowLevelBufferHandler` проверяет входные данные и выбрасывает исключение в случае ошибки
 - тип `BufferId` теперь занимает 4 байта
 - класс SharedBufferReader теперь умеет читать буфер (значения + признак достоверности) с помощью метода getBuffer. Чтение буфера длиной 65536 значений происходит за 0.6 мс.
 - поправлены тесты. Написаны новые

+++ src/lib/SharedBufferGlobal.h
@@ -3,7 +3,7 @@
 #include <QString>
 
 typedef quint32 BufferPos;
-typedef quint16 BufferId;
+typedef quint32 BufferId;
 typedef float SignalValue;
 typedef quint64 TimeStamp;
 typedef quint16 QualityCode;
